upnpevents: don't add to event engine a socket that failed connect(2)
The change 2c66335 fixed the bug when we were not adding a socket to
the event engine if connect(2) returned EINPROGRESS, but regressed to
leaving it in the event engine in case of other errors.

While here slighly streamline upnp_event_create_notify() so that it
doesn't link failed obj to the notify list.  In case of any error the
object is freed immediately.

NB: it could be that problem would expose itself only with kqueue(2),
which utilizes EV_FLAG_CLOSING and relies on close(2) happening in
tandem with event_module.del().