Make logging through syslog() a run-time option on FreeBSD
This feature is only compiled in on FreeBSD, if the #ifdef __FreeBSD__
clauses are kept. But it should apply to all systems that do not use
systemd as well, if the definition of SYSLOG_MASK to a non-zero value
is made unconditional (except on Linux, since I'm repurposing the -S
option, which may be a debatable decision).

This change allows the following patch to be applied to the service
script:

diff --git a/net/minidlna/files/minidlna.in b/net/minidlna/files/minidlna.in
index 4329e9c00927..a62cbd4c6632 100644
--- a/net/minidlna/files/minidlna.in
+++ b/net/minidlna/files/minidlna.in
@@ -19,11 +19,7 @@ minidlna_enable=${minidlna_enable-"NO"}
 minidlna_uid=${minidlna_uid-"%%USER%%"}

 command=%%PREFIX%%/sbin/minidlnad
-pidfile="/var/run/minidlna/minidlna.pid"
-command_args="-P $pidfile -u $minidlna_uid -f %%PREFIX%%/etc/minidlna.conf"
-
-start_precmd="install -d -o $minidlna_uid ${pidfile%/*} /var/db/minidlna"
-stop_postcmd="rm -f $pidfile"
+command_args="-S -u $minidlna_uid -f %%PREFIX%%/etc/minidlna.conf"

 extra_commands=rescan
 rescan_cmd="$command $command_args -R"

There is no need for a PID file, if logging is through syslog().