sendfile: fix FreeBSD shim to return bytes sent
This shim functions exist to bring sendfile API on different platforms
closer to Linux sendfile.  On Linux sendfile reports number of bytes sent
in case of success, and on FreeBSD it returns 0.

This becomes a bug only when a client sends open end request for a file
that is bigger than MAX_BUFFER_SIZE (which apparently is INT_MAX).  So the
shim function successfully sends 2 Gb and returns 0, after that the
upnphttp.c:send_file() just stops serving data.

This is a minimal fix.  A better refactoring would be to return 64-bit
value from the shim functions, just like Linux syscall does and remove
this MAX_BUFFER_SIZE, which apparently is INT_MAX, cause this artificial
limitation isn't needed.