Remove indentation from Makefile.am
On my FreeBSD system, the indentation of the "if HAVE_KQUEUE/endif"
block added in commit 1a2dfcc leads to a configure failure of the port
at the "depfiles" stage.

This commit removes the indentation, which fixes the problem.

See also https://bugs.freebsd.org/bugzilla/show_bug.cgi?id=288425,
which I opened before realizing that the maintainer of this github
repo is also the FreeBSD port maintainer.